+++ src/App.js
@@ -21,11 +21,11 @@ import '@fortawesome/fontawesome-free/css/all.min.css'
 import StoreAddPage from 'pages/StoreAddPage/StoreAddPage.jsx'
 import StoreList from 'pages/StoreList/StoreList.jsx' // 추가된 StoreList 컴포넌트
 import Mymaps from 'pages/My_Maps/Mymaps.jsx'
-import Review from 'pages/MyDining/Review/Review'
-import ReviewDetail from 'pages/MyDining/ReviewDetail/ReviewDetail'
 import Signup from './pages/Login/Signup/Signup.jsx'
 import RegisterType from 'pages/Login/Signup/RegisterType'
 import OwnnerSignup from 'pages/Login/Signup/OwnnerSignup/index.jsx'
+import ReviewModal from 'pages/MyDining/ReviewModal/ReviewModal.jsx'
+import ReviewDetailModal from 'pages/MyDining/ReviewDetailModal/ReviewDetaiModall.jsx'
 
 // 뭐징 17:06
 
@@ -43,8 +43,8 @@ const AppContent = () => {
                 <Route path="/SignType" element={<RegisterType />} />
                 <Route path="/OwnnerSignUp" element={<OwnnerSignup />} />
                 <Route path="/login" element={<Login />} />
-                <Route path='/mypage' element={<Mypage />} />
-                <Route path='/signup' element={<Signup />} />
+                <Route path="/mypage" element={<Mypage />} />
+                <Route path="/signup" element={<Signup />} />
                 <Route path="/maps" element={<Mymaps />} />
                 <Route path="/store/:storeId" element={<StoreDetail />} />
                 <Route path="/alarm" element={<Alam />} />
@@ -63,10 +63,10 @@ const AppContent = () => {
                 <Route path="/reserveDetail" element={<ReserveDetail />} />
                 <Route path="/storeadd" element={<StoreAddPage />} />
                 <Route path="/myDining" element={<MyDining />} />
-                <Route path="/Review" element={<Review />} />
+                <Route path="/ReviewModal" element={<ReviewModal />} />
                 <Route
-                    path="/ReviewDetail/:reservationId"
-                    element={<ReviewDetail />}
+                    path="/ReviewDetailModal/:reservationId"
+                    element={<ReviewDetailModal />}
                 />
                 <Route path="/storeList/:category" element={<StoreList />} />{' '}
                 {/* 추가된 StoreList 경로 */}
