--- src/pages/MyDining/Review/Review.module.css
@@ -1,49 +1,0 @@
-.reviewContainer {
-    padding: 20px;
-    max-width: 600px;
-    margin: 0 auto;
-    background-color: #f9f9f9;
-    border-radius: 8px;
-    box-shadow: 0 4px 8px rgba(0, 0, 0, 0.1);
-}
-
-.title {
-    font-size: 24px;
-    font-weight: bold;
-    margin-bottom: 20px;
-}
-
-.ratingContainer {
-    margin: 10px 0;
-    font-size: 24px;
-}
-
-.star {
-    font-size: 30px;
-    margin-right: 5px;
-}
-
-.reviewText {
-    width: 100%;
-    height: 100px;
-    padding: 10px;
-    margin: 10px 0;
-    border: 1px solid #ccc;
-    border-radius: 4px;
-    font-size: 16px;
-    resize: none;
-}
-
-.submitButton {
-    background-color: #4CAF50;
-    color: white;
-    padding: 10px 20px;
-    border: none;
-    border-radius: 4px;
-    cursor: pointer;
-    font-size: 16px;
-}
-
-.submitButton:disabled {
-    background-color: #ccc;
-}
