--- src/pages/MyDining/ReviewDetail/ReviewDetail.module.css
@@ -1,102 +1,0 @@
-.reviewContainer,
-.reviewDetailContainer {
-    max-width: 600px;
-    margin: 0 auto;
-    padding: 20px;
-    border-radius: 8px;
-    background-color: #f9f9f9;
-    box-shadow: 0 0 10px rgba(0, 0, 0, 0.1);
-    height: 400px;
-}
-
-.title {
-    font-size: 1.5rem;
-    text-align: center;
-    margin-bottom: 20px;
-}
-
-.reviewInfo {
-    margin-bottom: 15px;
-}
-
-.reviewText {
-    white-space: pre-wrap;
-    line-height: 1.5;
-}
-
-.textAreaEdit {
-    width: 100%;
-    /* div와 동일한 너비로 설정 */
-    height: auto;
-    /* 자동으로 높이를 조절하거나 원하는 높이 설정 */
-    font-size: 1rem;
-    /* 원래 텍스트 크기와 동일하게 */
-    padding: 10px;
-    /* div와 동일한 padding */
-    line-height: 1.5;
-    /* 텍스트 줄 간격 */
-    border: 1px solid #ccc;
-    /* 테두리 설정 */
-    border-radius: 4px;
-    /* 부드러운 모서리 */
-    box-sizing: border-box;
-    /* padding을 포함한 전체 크기 조절 */
-    background-color: #f9f9f9;
-    /* 배경색 div와 동일하게 설정 */
-    white-space: pre-wrap;
-    /* 기존 텍스트 스타일과 동일하게 줄바꿈 */
-    resize: none;
-    /* 크기 조절 불가하게 설정 (필요할 경우) */
-}
-
-.ratingContainer {
-    display: flex;
-    gap: 5px;
-}
-
-.star {
-    font-size: 24px;
-}
-
-.submitButton {
-    background-color: #007BFF;
-    color: white;
-    padding: 10px 20px;
-    border: none;
-    border-radius: 5px;
-    cursor: pointer;
-}
-
-.submitButton:disabled {
-    background-color: #ccc;
-}
-
-.imagePlaceholder {
-    width: 100px;
-    height: 100px;
-    background-color: #ddd;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    margin-right: 20px;
-}
-
-.editButton {
-    background-color: #4CAF50;
-    color: white;
-    padding: 10px 20px;
-    border: none;
-    border-radius: 4px;
-    cursor: pointer;
-    font-size: 16px;
-}
-
-.deleteButton {
-    background-color: #ff6b00;
-    color: white;
-    padding: 10px 20px;
-    border: none;
-    border-radius: 4px;
-    cursor: pointer;
-    font-size: 16px;
-}
